movie_reviewer_api/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==26.0.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
